+++ database.go
@@ -6,14 +6,14 @@ import (
 )
 
 type Order struct {
-	OID            int    `gorm:"primary_key;column:OID"`
-	Name           string `gorm:"column:Name"`
-	Barcode        string `gorm:"column:Barcode"`
-	ProductID      int    `gorm:"column:ProductID"`
-	OrderStatusID  int    `gorm:"column:OrderStatusID"`
-	CountRequested int    `gorm:"column:CountRequested"`
-	WorkplaceID    int    `gorm:"column:WorkplaceID"`
-	Cavity         int    `gorm:"column:Cavity"`
+	OID            int           `gorm:"primary_key;column:OID"`
+	Name           string        `gorm:"column:Name"`
+	Barcode        string        `gorm:"column:Barcode"`
+	ProductID      int           `gorm:"column:ProductID"`
+	OrderStatusID  int           `gorm:"column:OrderStatusID"`
+	CountRequested int           `gorm:"column:CountRequested"`
+	WorkplaceID    sql.NullInt32 `gorm:"column:WorkplaceID"`
+	Cavity         int           `gorm:"column:Cavity"`
 }
 
 func (Order) TableName() string {
@@ -58,17 +58,28 @@ func (TerminalInputIdle) TableName() string {
 }
 
 type TerminalInputLogin struct {
-	OID      int       `gorm:"primary_key;column:OID"`
-	DTS      time.Time `gorm:"column:DTS"`
-	DTE      time.Time `gorm:"column:DTE"`
-	UserID   int       `gorm:"column:UserID"`
-	Interval float32   `gorm:"column:Interval"`
-	DeviceID int       `gorm:"column:DeviceID"`
-	Note     string    `gorm:"column:Note"`
+	OID      int          `gorm:"primary_key;column:OID"`
+	DTS      time.Time    `gorm:"column:DTS"`
+	DTE      sql.NullTime `gorm:"column:DTE"`
+	UserID   int          `gorm:"column:UserID"`
+	Interval float32      `gorm:"column:Interval"`
+	DeviceID int          `gorm:"column:DeviceID"`
+	Note     string       `gorm:"column:Note"`
 }
 
 func (TerminalInputLogin) TableName() string {
 	return "terminal_input_login"
+}
+
+type WorkplaceMode struct {
+	OID                 int `gorm:"primary_key;column:OID"`
+	WorkplaceID         int `gorm:"column:WorkplaceID"`
+	IdleFromTime        int `gorm:"column:IdleFromTime"`
+	WorkplaceModeTypeID int `gorm:"column:WorkplaceModeTypeID"`
+}
+
+func (WorkplaceMode) TableName() string {
+	return "workplace_mode"
 }
 
 type TerminalInputOrder struct {
@@ -144,4 +155,41 @@ type WorkplaceDivision struct {
 
 func (WorkplaceDivision) TableName() string {
 	return "workplace_division"
+}
+
+type SkZapsiVp struct {
+	RID         int     `gorm:"primary_key;column:RID"`
+	VP          string  `gorm:"column:VP"`
+	VPexp       string  `gorm:"column:VPexp"`
+	Stredisko   string  `gorm:"column:Stredisko"`
+	Mnoz_PL     float32 `gorm:"column:Mnoz_PL"`
+	Zkr         string  `gorm:"column:Zkr"`
+	Naz         string  `gorm:"column:Naz"`
+	MaterialZkr string  `gorm:"column:MaterialZkr"`
+	MaterialNaz string  `gorm:"column:MaterialNaz"`
+	MaterialEAN string  `gorm:"column:MaterialEAN"`
+	MaterialM2  float32 `gorm:"column:MaterialM2"`
+	MaterialBM  float32 `gorm:"column:MaterialBM"`
+	Priprava    float32 `gorm:"column:Priprava"`
+	Tisk        float32 `gorm:"column:Tisk"`
+	Myti        float32 `gorm:"column:Myti"`
+}
+
+func (SkZapsiVp) TableName() string {
+	return "SK_ZAPSI_VP"
+}
+
+type ZapsiK2 struct {
+	Cas        time.Time     `gorm:"column:cas"`
+	Typ        int           `gorm:"column:typ"`
+	Data       string        `gorm:"column:data"`
+	Zprac      int           `gorm:"column:zprac"`
+	Cas_zprac  sql.NullTime  `gorm:"column:cas_zprac"`
+	Error      sql.NullInt32 `gorm:"column:error"`
+	CasTrigger sql.NullTime  `gorm:"column:castrigger"`
+	Idzaznamu  sql.NullInt32 `gorm:"primary_key;column:id_zaznamu"`
+}
+
+func (ZapsiK2) TableName() string {
+	return "ZAPSI_K2"
 }