+++ js/idle_running.js
@@ -16,7 +16,14 @@ idleEnd.addEventListener("click", endIdle)
 
 function endIdle() {
     console.log("clicked")
-    let data = {Data: "K219"};
+    let data = {
+        Type: "idle",
+        Code: "K119",
+        WorkplaceCode: workplace.textContent,
+        UserId: sessionStorage.getItem("UserId"),
+        OrderBarcode: sessionStorage.getItem("Order"),
+        IdleId: sessionStorage.getItem("IdleId"),
+    };
     fetch("/save_code", {
         method: "POST",
         body: JSON.stringify(data)