+++ main.go
@@ -15,7 +15,7 @@ const version = "2020.4.1.2"
 const serviceName = "SK Label Cutting Webservice"
 const serviceDescription = "Web Service for terminals for cutting workplaces"
 const zapsiDatabaseConnection = "zapsi_uzivatel:zapsi@tcp(localhost:3306)/zapsi2?charset=utf8mb4&parseTime=True&loc=Local"
-const skLabelDatabaseConnection = "user=postgres password=Zps05..... dbname=version3 host=database port=5432 sslmode=disable"
+const skLabelDatabaseConnection = "sqlserver://zapsi:DSgEEmPNxCwgTJjsd2uR@10.3.1.3/K2_SKLABEL?database=SK_Zapsi_VP"
 
 var devicesMap map[string]string